+++ src/index.html
@@ -172,6 +172,23 @@
 
 
     <!-- Statistics page -->
+    <main class="stats-page">
+    <table>
+        <thead>
+        <tr>
+          <td><strong>Country</strong></td>
+          <td><strong>Gold</strong></td>
+          <td><strong>Silver</strong></td>
+          <td><strong>Bronze</strong></td>
+          <td><strong>Total</strong></td>
+        </tr>
+        </thead>
+        <tbody id="bodytable">
+
+        </tbody>
+         
+    </table>
+    </main>
 
     <footer class="footer"> © 2021 T, C and MC - All Rights Reserved </footer>
