--- src/statistics.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <title>STATISTICS OLYMPICS RIO 2016</title>
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Montserrat&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="style.css" />
-
-</head>
-<body class="body">
-    <header class="header">
-        <div class="header__logo-area">
-            <img class="header__logo-area__logo" src="img/logo-olimpiadas.png">
-        </div>
-
-        <div class="header__nav"> 
-            <a class="header__nav__link" id="home-button" href="#HOME">HOME</a>
-            <a class="header__nav__link" id="champs-button" href="#CHAMPIONS">CHAMPIONS</a>
-            <a class="header__nav__link" id="stats-button" href="#STATISTICS">STATISTICS</a>
-        </div>
-        <div class="header__nav-mobile">
-            <div class="hambuger-menu">
-                <div class="bar1"></div>
-                <div class="bar2"></div>
-                <div class="bar3"></div>
-            </div>
-        </div>
-    </header>
-    <table>
-        <thead>
-        <tr>
-          <td><strong>Country</strong></td>
-          <td><strong>Gold</strong></td>
-          <td><strong>Silver</strong></td>
-          <td><strong>Bronze</strong></td>
-          <td><strong>Total</strong></td>
-        </tr>
-        </thead>
-        <tbody id="bodytable">
-
-        </tbody>
-         
-    </table>
-    
-    <footer class="footer"> © 2021 T, C and MC - All Rights Reserved </footer>
- 
-    <script src="statistics.js" type="module"></script>
-    
-</body>
-</html>
